Rename and improve MD5 digest function
