+++ include/appimage/appimage.h
@@ -133,8 +133,10 @@ ssize_t appimage_get_elf_size(const char* fname);
  * You need to allocate a char array of at least 16 bytes (128 bit) and pass a reference to it as digest parameter.
  * The function will set it to the raw digest, without any kind of termination. Please use appimage_hexlify() if you
  * need a textual representation.
+ *
+ * Please beware that this calculation is only available for type 2 AppImages.
  */
-bool appimage_digest_md5(const char* fname, char* digest);
+bool appimage_type2_digest_md5(const char* fname, char* digest);
 
 /*
  * Creates hexadecimal representation of a byte array. Allocates a new char array (string) with the correct size that
